+++ OneBlink.SDK/models/TilesFormsApp.cs
@@ -14,5 +14,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public override string type
+        {
+            get;
+        }
     }
 }